Move StatsCollector behind an EngineStats package trait
Extracts StatsCollector and the _Stderr output stream into their own file,
gated by a new (off-by-default) EngineStats trait. _Stderr writes through the
platform libc directly, so _CWasmKit.h no longer needs stdio/stdlib or the
wasmkit_fwrite_stderr helper. Also fixes stale variable reference in the
stats tracking call, which only compiled when stats were disabled.
