+++ Sources/_CWasmKit/include/_CWasmKit.h
@@ -4,8 +4,6 @@
 #include <stdatomic.h>
 #include <stddef.h>
 #include <stdint.h>
-#include <stdio.h>
-#include <stdlib.h>
 
 #include "Platform.h"
 
@@ -93,10 +91,6 @@ static inline void wasmkit_tc_start(
   exec(sp, pc, md, ms, state);
 }
 #endif
-
-static inline void wasmkit_fwrite_stderr(const char *_Nonnull str, size_t len) {
-  fwrite(str, 1, len, stderr);
-}
 
 int wasmkit_address_sanitizer_enabled(void);
